refactor: refactor: Name bootstrap module search_paths
